==> analysis/03_pp_qc.Rmd <==
---
title: "Preprocessing 03: Quality control"
author: "Kevin Thomas"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  rmarkdown::github_document:
    html_preview: false
    toc: true
    toc_depth: 3
---

```{css, echo=FALSE}
    body .main-container {
      max-width: 1500px !important;
      width: 1500px !important;
    }
    body {
      max-width: 1500px !important;
    }
```

```{r setup, include=FALSE}
# Set options
here::i_am("analysis/03_pp_qc.Rmd")
options(width = 1200)
knitr::opts_chunk$set(echo = TRUE, fig.width = 12, fig.height = 12)
options(future.globals.maxSize = +Inf)

# Attach libraries
library(here)
library(Seurat)
library(BiocParallel)
library(BiocNeighbors)
library(future)
library(dplyr)
library(stringr)
library(Matrix)
library(magrittr)
library(devtools)

# Parallelization
BPPARAM <- SnowParam(workers = ceiling(0.5 * parallel::detectCores()), type = "SOCK")
BNPARAM <- HnswParam()
register(BPPARAM)
plan(strategy = multisession, workers = parallel::detectCores())

# Define custom functions
source(here("R", "functions.R"))

# Define variables
primary_species_prefix <- "GRCh38_"
secondary_species_prefix <- "mm10___"
IG_genes <- c(
  "IGHA1",
  "IGHA2",
  "IGHG1",
  "IGHG2",
  "IGHG3",
  "IGHG4",
  "IGHGP",
  "IGHD",
  "IGHE",
  "IGHM",
  "IGLC1",
  "IGLC2",
  "IGLC3",
  "IGLC4",
  "IGLC5",
  "IGLC6",
  "IGLC7",
  "IGKC"
)
```

## Quality control of cells for initial analysis

```{r load pre_qc and gem_classification}
pre_qc <- readRDS(file = here("demo_data", "pre_qc.RDS"))
gem_classification <- readRDS(file = here("demo_data", "gem_classification.RDS"))
```

```{r separate mouse vs human}
# First, separate clear mouse and human cells
human <- subset(
  x = pre_qc,
  cells = gem_classification[which(gem_classification$call == str_remove_all(primary_species_prefix, "_")), ] |>
    mutate(cell_name = paste0(run, "_", cells)) |>
    pull(cell_name)
) |>
  PurgeXenoGenes(
    primary_species_prefix = str_replace_all(primary_species_prefix, "_", "-"),
    secondary_species_prefix = str_replace_all(secondary_species_prefix, "_", "-")
  )

mouse <- subset(
  x = pre_qc,
  cells = gem_classification[which(gem_classification$call == str_remove_all(secondary_species_prefix, "_")), ] |>
    mutate(cell_name = paste0(run, "_", cells)) |>
    pull(cell_name)
) |>
  PurgeXenoGenes(
    primary_species_prefix = str_replace_all(secondary_species_prefix, "_", "-"),
    secondary_species_prefix = str_replace_all(primary_species_prefix, "_", "-")
  )
```

```{r filter on singlets}
# Filter on singlets only
human_singlets <- if ("HTO" %in% Seurat::Assays(human)) {
  subset(x = human, `final.HTO.ID` %nin% c("Negative", "Doublet"))
} else {
  pre_qc[, colnames(pre_qc) %in% gem_classification[gem_classification$call == str_remove(primary_species_prefix, "_"), ]$sample_barcode] |>
    PurgeXenoGenes(
      primary_species_prefix = str_replace_all(primary_species_prefix, "_", "-"),
      secondary_species_prefix = str_replace_all(secondary_species_prefix, "_", "-")
    )
}
# Free memory
rm(pre_qc, human, mouse)
gc()
```

```{r qc cutoffs}
# Perform QC by setting global cutoffs
qc_object <- PercentageFeatureSet(
  object = human_singlets,
  pattern = "^MT-",
  col.name = "percent_mt"
) %>%
  PercentageFeatureSet(
    pattern = "^HB[ABDEGMQZ].?$",
    col.name = "percent_hemo"
  ) %>%
  subset(
    features = c(
      paste0("rna_", names(which(rowSums(.[["RNA"]]@counts >= 1) >= 3))),
      paste0("cite_", rownames(.[["CITE"]])),
      paste0("hto_", rownames(.[["HTO"]]))
    )
  ) %>%
  # Remove the low feature, high mito, and high hemo count cells
  subset(
    `nFeature_RNA` >= 40 &
      `percent_mt` <= 25 &
      `percent_hemo` <= 1.5
  ) |>
  # Score Ig gene differences
  ScoreIgDiffs(
    assay = "RNA",
    ig.hc = IG_genes[1:10],
    ig.lc = IG_genes[11:18],
    conserve.memory = TRUE
  )
```

```{r save data}
# Save data for next step
saveRDS(object = qc_object, file = here("demo_data", "qc_object.RDS"))
```

```{r}
session_info()
```
