==> analysis/04_analysis.Rmd <==
---
title: "Analysis 04: Analysis and clustering"
author: "Kevin Thomas"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  rmarkdown::github_document:
    html_preview: false
    toc: true
    toc_depth: 3
---

```{css, echo=FALSE}
    body .main-container {
      max-width: 1500px !important;
      width: 1500px !important;
    }
    body {
      max-width: 1500px !important;
    }
```

```{r setup, include=FALSE}
# Set options
here::i_am("analysis/04_analysis.Rmd")
options(width = 1200)
reticulate::use_condaenv(condaenv = "r-reticulate", required = TRUE, conda = "~/.local/share/r-miniconda/bin/conda")
knitr::opts_chunk$set(echo = TRUE, fig.width = 12, fig.height = 12)
options(future.globals.maxSize = +Inf)

# Attach libraries
library(here)
library(Seurat)
library(BiocParallel)
library(BiocNeighbors)
library(future)
library(magrittr)
library(harmony)
library(purrr)
library(igraph)
library(leiden)
library(future.apply)
library(stringr)
library(devtools)

# Parallelization
BPPARAM <- SnowParam(workers = ceiling(0.5 * parallel::detectCores()), type = "SOCK")
BNPARAM <- HnswParam()
register(BPPARAM)

# Define custom functions
source(here("R", "functions.R"))
```

## Initial analysis of QC'd cells

```{r load qc object}
qc_object <- readRDS(file = here("demo_data", "qc_object.RDS"))
```

### Normalization

```{r normalization}
# Reduce available cores to ensure sufficient memory per core
plan(strategy = multisession, workers = ceiling(0.25 * parallel::detectCores()))
# SCTransform normalization. Regress out mitochondrial content,
# hemoglobin content, differences in Ig subclass, and cell cycle phase
normalized_object <- SCTransform(
  object = qc_object,
  assay = "RNA",
  vars.to.regress =
    c(
      "percent_mt",
      "percent_hemo",
      grep(
        pattern = "[.]diff$",
        x = names(qc_object[[]]),
        value = TRUE
      )
    ),
  method = "glmGamPoi"
)
# Reset paralellization
plan(strategy = multisession, workers = parallel::detectCores())
```

## Harmony on PC
```{r harmonization}
# Clear and free up memory
rm(qc_object)
gc()
# Run PCA and harmonize on normalized gene expression and CITE expression
harmonized_object <- normalized_object |>
  RunPCA() |>
  FindPCAElbow(
    assay = "SCT",
    perform_new_pca = FALSE,
    elbow_th = 0.025
  ) |>
  RunHarmony(
    group.by.vars = c("run"),
    assay.use = "SCT",
    dims.use = 1:50
  ) |>
  ScaleData(
    assay = "CITE"
  ) |>
  `DefaultAssay<-`(
    value = "CITE"
  ) %>%
  RunPCA(
    features = rownames(.)[rownames(.) != "isotype-control"],
    reduction.name = "pca_cite",
    reduction.key = "citePCA_"
  ) |>
  FindPCAElbow(
    assay = "CITE",
    reduction = "pca_cite",
    perform_new_pca = FALSE,
    elbow_th = 0.025
  ) |>
  RunHarmony(
    group.by.vars = c("run"),
    assay.use = "CITE",
    reduction = "pca_cite",
    dims.use = 1:50,
    reduction.save = "harmony_cite"
  ) |>
  `DefaultAssay<-`(
    value = "SCT"
  )
```

## Weighted nearest neighbors (wnn) analysis
```{r wnn}
# Clear and free up memory
rm(normalized_object)
gc()
# Find weighted nearest neighbors with RNA and CITE
wnn_obj <-
  FindMultiModalNeighbors(
    object = harmonized_object,
    reduction.list = list("harmony", "harmony_cite"),
    dims.list =
      list(
        seq(harmonized_object@misc$max_pca_dim),
        seq(harmonized_object@misc$max_pca_cite_dim)
      ),
    modality.weight.name = c("SCT.weight", "CITE.weight")
  ) |>
  # Make UMAP reduction on wnn graph
  RunUMAP(
    nn.name = "weighted.nn",
    reduction.name = "umap_wnn",
    reduction.key = "wnnUMAP_"
  )
```

## Clustering
```{r cluster}
# Clear and free up memory
rm(harmonized_object)
gc()
# Leiden clustering at multiple resolutions
wnn_clust_obj <- wnn_obj %>%
  (function(seurat_obj) {
    obj <- as(object = seurat_obj@graphs$wsnn, Class = "dgCMatrix")
    input <- graph_from_adjacency_matrix(adjmatrix = obj, weighted = TRUE)
    clustering.results <- future_lapply(
      X = seq(0.2, 1.2, 0.2),
      FUN = function(r) {
        ids <- leiden(
          object = input,
          resolution_parameter = r,
          seed = 42,
          weights = NULL
        )
        names(x = ids) <- colnames(x = obj)
        ids <- GroupSmallClusters(ids = ids, SNN = obj, threshold = 20, verbose = TRUE)
        results <- list(factor(x = ids, levels = str_sort(unique(ids), numeric = TRUE)))
        names(x = results) <- paste0("res.", r)
        return(results)
      }
    )
    clustering.results <- as.data.frame(x = clustering.results)
    seurat_obj@meta.data[, paste0("wsnn.res.", seq(0.2, 1.2, 0.2))] <- clustering.results
    seurat_obj
  })
```

```{r save data}
# Save data for next step
saveRDS(object = wnn_clust_obj, file = here("demo_data", "wnn_clust_obj.RDS"))
```

```{r}
session_info()
```
