==> analysis/05_plotting.Rmd <==
---
title: "Analysis 05: Plotting demonstration"
author: "Kevin Thomas"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  rmarkdown::github_document:
    html_preview: false
    toc: true
    toc_depth: 3
---

```{css, echo=FALSE}
    body .main-container {
      max-width: 1500px !important;
      width: 1500px !important;
    }
    body {
      max-width: 1500px !important;
    }
```

```{r setup, include=FALSE}
# Set options
here::i_am("analysis/05_plotting.Rmd")
options(width = 1200)
knitr::opts_chunk$set(echo = TRUE, fig.width = 12, fig.height = 12)
reticulate::use_condaenv(condaenv = "r-reticulate", required = TRUE, conda = "~/.local/share/r-miniconda/bin/conda")
options(future.globals.maxSize = +Inf)

# Attach libraries
library(here)
library(Seurat)
library(BiocParallel)
library(BiocNeighbors)
library(future)
library(cowplot)
library(ggplot2)
library(magrittr)
library(purrr)
library(dplyr)
library(tibble)
library(presto)
library(ggrepel)
library(devtools)

# Parallelization
BPPARAM <- SnowParam(workers = ceiling(0.5 * parallel::detectCores()), type = "SOCK")
BNPARAM <- HnswParam()
register(BPPARAM)
plan(strategy = multisession, workers = parallel::detectCores())

# Define custom functions
source(file = here("R", "functions.R"))

# Set plot theme for any plots
theme_set(cowplot::theme_cowplot())
```

## Plotting of initial analysis

```{r load qc object}
wnn_clust_obj <- readRDS(file = here("demo_data", "wnn_clust_obj.RDS"))
```

### All resolutions
```{r resolution, fig.height=11.25, fig.width=20}
# Plot
map(
  .x = names(select(wnn_clust_obj@meta.data, starts_with(match = "wsnn."))),
  .f = function(x) {
    DimPlot(object = wnn_clust_obj, reduction = "umap_wnn", group.by = x, label = TRUE) +
      NoLegend() +
      labs(title = x) +
      coord_fixed()
  }
) %>% plot_grid(plotlist = ., ncol = 3)
```

### Top 5 Gene markers at each resolution
```{r top5 gene, fig.height=19.526, fig.width=40}
map(
  .x = names(select(wnn_clust_obj@meta.data, starts_with(match = "wsnn."))),
  .f = function(x) {
    # x <- names(select(wnn_clust_obj@meta.data, starts_with(match = "wsnn.")))[1]
    top5_gene <- wilcoxauc(
      X = wnn_clust_obj,
      seurat_assay = "SCT",
      group_by = x
    ) |>
      as_tibble() |>
      select(
        feature,
        cluster = group,
        p_val = pval,
        p_val_adj = padj,
        avg_logFC = logFC,
        pct.1 = pct_in,
        pct.2 = pct_out,
        auc
      ) |>
      filter(p_val_adj < 0.05) |>
      mutate(cluster = factor(cluster, levels = levels(wnn_clust_obj@meta.data[, x]))) %>%
      split.data.frame(use_series(., cluster)) |>
      map(filter, avg_logFC > 0, pct.2 < 0.5) |>
      map(arrange, cluster, desc(auc), p_val_adj) |>
      map(top_n, 5, auc) |>
      map(select, feature) |>
      map(deframe) |>
      map(paste, collapse = "\n")

    DimPlot(object = wnn_clust_obj, reduction = "umap_wnn", group.by = x) +
      geom_label_repel(
        data = DimPlot(wnn_clust_obj, reduction = "umap_wnn", group.by = x)[[1]]$data |>
          group_by_at(x) |>
          summarize_all(median),
        aes(wnnUMAP_1, wnnUMAP_2, label = paste0(get(x), "\n", top5_gene), fill = get(x))
      ) +
      NoLegend() +
      labs(title = x) +
      coord_fixed()
  }
) %>% plot_grid(plotlist = ., ncol = 3)
```

### Top 5 CITE markers at each resolution
```{r top5 cite, fig.height=19.526, fig.width=40}
map(
  .x = names(select(wnn_clust_obj@meta.data, starts_with(match = "wsnn."))),
  .f = function(x) {
    # x <- names(select(wnn_clust_obj@meta.data, starts_with(match = "wsnn.")))[1]
    top5_gene <- wilcoxauc(
      X = wnn_clust_obj,
      seurat_assay = "CITE",
      group_by = x
    ) |>
      as_tibble() |>
      select(
        feature,
        cluster = group,
        p_val = pval,
        p_val_adj = padj,
        avg_logFC = logFC,
        pct.1 = pct_in,
        pct.2 = pct_out,
        auc
      ) |>
      filter(p_val_adj < 0.05) |>
      mutate(cluster = factor(cluster, levels = levels(wnn_clust_obj@meta.data[, x]))) %>%
      split.data.frame(use_series(., cluster)) %>%
      map(filter, avg_logFC > 0) |>
      map(arrange, cluster, desc(avg_logFC), p_val_adj) |>
      map(top_n, 5, avg_logFC) |>
      # map(top_n, 5, auc) |>
      map(select, feature) |>
      map(deframe) |>
      map(paste, collapse = "\n")

    DimPlot(object = wnn_clust_obj, reduction = "umap_wnn", group.by = x) +
      geom_label_repel(
        data = DimPlot(wnn_clust_obj, reduction = "umap_wnn", group.by = x)[[1]]$data |>
          group_by_at(x) |>
          summarize_all(median),
        aes(wnnUMAP_1, wnnUMAP_2, label = paste0(get(x), "\n", top5_gene), fill = get(x))
      ) +
      NoLegend() +
      labs(title = x) +
      coord_fixed()
  }
) %>% plot_grid(plotlist = ., ncol = 3)
```

## Gene markers of major cell types
```{r genes, fig.height=11.25, fig.width=20}
FeaturePlot(
  `DefaultAssay<-`(object = wnn_clust_obj, value = "SCT"),
  features = c(
    "CD74", "MS4A1", "JCHAIN", "MZB1", "ZNF385D", "SOX4", "HLA-DPB1",
    "FCER1A", "S100A8", "LYZ", "GNLY", "NKG7", "TCF4", "IRF7", "CD3D", "IL32"
  ),
  reduction = "umap_wnn",
  coord.fixed = T,
  cols = c("lightgrey", "blue2"),
  order = T,
  combine = FALSE
) |>
  map(add, labs(x = NULL, y = NULL)) %>%
  plot_grid(plotlist = ., align = "hv")
```

## CITE markers of major cell types
```{r CITE, fig.height=11.25, fig.width=20}
FeaturePlot(
  `DefaultAssay<-`(object = wnn_clust_obj, value = "CITE"),
  features = c(
    "CD19", "CD21", "CD38", "CD27", "CD34", "CD117", "CD1c", "HLA-DR", "CD11c",
    "CD14", "CD56", "CD16", "CD123", "CD4", "CD8a", "CD127"
  ),
  reduction = "umap_wnn",
  coord.fixed = T,
  cols = c("lightgrey", "red2"),
  min.cutoff = "q05",
  max.cutoff = "q95",
  combine = FALSE
) |>
  map(add, labs(x = NULL, y = NULL)) %>%
  plot_grid(plotlist = ., align = "hv")
```

```{r}
session_info()
```
